Struct blender_armature::BlenderArmature

pub struct BlenderArmature {
    pub joint_index: HashMap<String, u8>,
    pub inverse_bind_poses: Vec<Bone>,
    pub actions: HashMap<String, Vec<Keyframe>>,
    // some fields omitted
}

All of the data about a Blender armature that we've exported from Blender. A BlenderArmature should have all of the data that you need to implement skeletal animation.

If you have other needs, such as a way to know the model space position of any bone at any time so that you can, say, render a baseball in on top of your hand bone.. Open an issue. (I plan to support this specific example in the future)

TODO: BlenderArmature.y_up() fixes the actions to be y up instead of z up

TODO: Replace String's with generics so that you can have type safety across different armature properties

Fields
joint_index: HashMap<String, u8>inverse_bind_poses: Vec<Bone>actions: HashMap<String, Vec<Keyframe>>
Implementations

impl BlenderArmature

pub fn matrix_to_dual_quat(bone: &Bone) -> Bone

Convert a matrix into a dual quaternion https://github.com/chinedufn/mat4-to-dual-quat/blob/master/src/mat4-to-dual-quat.js Note that we use w, x, y, z and not x, y, z, w for our quaternion representation

pub fn dual_quat_to_matrix(bone: &Bone) -> Bone

https://github.com/chinedufn/dual-quat-to-mat4/blob/master/src/dual-quat-to-mat4.js

impl BlenderArmature

pub fn change_coordinate_system(&mut self, system: CoordinateSystem)

Shift around the data in the armature to a new coordinate system.

For example, if the armature was previously Z up and we're switching to Y up

the new +Y axis would be the old +Z axis

the new +Z axis would be the old -Y axis

impl BlenderArmature

pub fn interpolate_bones(
&self,
opts: InterpolationSettings
) -> BTreeMap<u8, Bone>

Interpolate in between the keyframes of your BlenderArmature. This is useful for skeletal animation.

We return a hashmap so that you can easily merge the results of interpolating different sets of bone groups.

Panics

We don't yet support interpolating matrix bones, so we panic if your bones aren't dual quaternions.

Panics if you pass in previous actions that do not have the exact same joint indices as your current action.

TODO

[ ] Return Result<HashMap<u8, Bone>, InterpolationError>

[ ] error if clock time is negative

impl BlenderArmature

pub fn bone_groups(&self) -> &HashMap<String, Vec<u8>>

Blender bone groups

Maps bone group name to a vector of the bones indices that are in that bone group.

let armature = create_blender_armature();

let joint_indices = armature.bone_groups().get("My bone group").unwrap();

let interpolate_opts = InterpolationSettings {
           current_time: 1.0,
           // FIXME: Base joint_indices on a property of `SkinnedMesh`
           joint_indices,
           blend_fn: None,
           current_action: &get_action(),
           previous_action: None,

};

let _bones = armature.interpolate_bones(interpolate_opts);

pub fn create_bone_group(&mut self, name: String, joint_indices: Vec<u8>)

Create a new bone group

impl BlenderArmature

pub fn apply_inverse_bind_poses(&mut self)

Iterate over all of the action bones and apply and multiply in the inverse bind pose.

TODO: another function to apply bind shape matrix? Most armatures seem to export an identity bind shape matrix but that might not be the same for every armature.

pub fn transpose_actions(&mut self)

Tranpose all of the bone matrices in our armature's action keyframes. Blender uses row major matrices, but OpenGL uses column major matrices so you'll usually want to transpose your matrices before using them.

impl BlenderArmature

pub fn matrices_to_dual_quats(&mut self)

Convert your action matrices into dual quaternions so that you can implement dual quaternion linear blending.